# openapi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "77-Bit API",
    "description": "An API for interacting with the 77-Bit L2 blockchain.",
    "termsOfService": "https://77-bit.com/terms",
    "license": {
      "name": "License",
      "url": "https://77-bit.com/license"
    },
    "version": "1.0",
    "x-logo": {
      "url": "https://pbs.twimg.com/profile_images/1642220143915696128/x3yCWwbE_400x400.jpg",
      "altText": "77-Bit logo"
    }
  },
  "servers": [
    {
      "url": "https://testnet-api.77-bit.com",
      "description": "Testnet server (Goerli)."
    },
    {
      "url": "https://api.77-bit.com",
      "description": "Mainnet server (Ethereum)."
    }
  ],
  "paths": {
    "/api/v1/assets": {
      "post": {
        "tags": [
          "Asset"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAssetModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "oauth2": [
              "write:assets"
            ]
          }
        ]
      },
      "get": {
        "tags": [
          "Asset"
        ],
        "summary": "Get All Assets",
        "description": "This endpoint fetches all assets in the system, with support for filters and pagination.",
        "operationId": "GetAllAssets",
        "parameters": [
          {
            "name": "asset_type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/AssetType"
            }
          },
          {
            "name": "page_number",
            "in": "query",
            "schema": {
              "title": "Page Number",
              "maximum": 2147483647,
              "minimum": 0,
              "type": "integer",
              "description": "The page number to retrieve.",
              "format": "int32"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "title": "Page Size",
              "maximum": 100,
              "minimum": 0,
              "type": "integer",
              "description": "The number of items to retrieve per page.",
              "format": "int32"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "schema": {
              "title": "Sort By",
              "type": "string",
              "description": "The field to sort the results by."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all assets in the system (paginated).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDtoPaginatedResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "oauth2": [
              "read:assets"
            ]
          }
        ]
      }
    },
    "/api/v1/assets/{assetId}": {
      "get": {
        "tags": [
          "Asset"
        ],
        "summary": "Get Asset",
        "description": "This endpoint fetches a specific asset by ID.",
        "operationId": "GetAsset",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "description": "The asset ID.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an asset in the system.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "oauth2": [
              "read:assets"
            ]
          }
        ]
      }
    },
...
    "/api/v1/transactions": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "Get All Transactions",
        "description": "This endpoint fetches all transactions submitted by the system, with support for filters and pagination.",
        "operationId": "GetAllTransactions",
        "parameters": [
          {
            "name": "transaction_status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TransactionStatus"
            }
          },
          {
            "name": "transaction_status_comparison",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/FilterOptions"
            }
          },
          {
            "name": "starkex_tx_id",
            "in": "query",
            "schema": {
              "title": "StarkEx Transaction ID",
              "type": "integer",
              "description": "The unique identifier of the StarkEx transaction.",
              "format": "integer"
            }
          },
          {
            "name": "starkex_tx_id_comparison",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/FilterOptions"
            }
          },
          {
            "name": "tx_type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/StarkExOperation"
            }
          },
          {
            "name": "tx_type_comparison",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/FilterOptions"
            }
          },
          {
            "name": "page_number",
            "in": "query",
            "schema": {
              "title": "Page Number",
              "maximum": 2147483647,
              "minimum": 0,
              "type": "integer",
              "description": "The page number to retrieve.",
              "format": "int32"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "title": "Page Size",
              "maximum": 100,
              "minimum": 0,
              "type": "integer",
              "description": "The number of items to retrieve per page.",
              "format": "int32"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "schema": {
              "title": "Sort By",
              "type": "string",
              "description": "The field to sort the results by."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all transactions submitted by the system (paginated).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDtoPaginatedResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "oauth2": [
              "read:transactions"
            ]
          }
        ]
      }
    },
    "/api/v1/transactions/{transactionId}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "Get Transaction",
        "description": "This endpoint fetches a specific transaction by ID.",
        "operationId": "GetTransaction",
        "parameters": [
          {
            "name": "transactionId",
            "in": "path",
            "description": "The transaction ID.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a transaction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "oauth2": [
              "read:transactions"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssetDto": {
        "type": "object",
        "properties": {
          "assetId": {
            "title": "Asset ID",
            "type": "string",
            "description": "The ID of the asset.",
            "format": "uuid"
          },
          "starkExType": {
            "title": "StarkEx Asset type",
            "type": "string",
            "description": "The StarkEx asset type.",
            "format": "hex",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/AssetType"
          },
          "address": {
            "title": "Asset Address",
            "type": "string",
            "description": "The ethereum address of the asset.",
            "format": "hex",
            "nullable": true
          },
          "name": {
            "title": "Asset Name",
            "type": "string",
            "description": "The name of the asset.",
            "format": "string",
            "nullable": true
          },
          "symbol": {
            "title": "Asset Symbol",
            "type": "string",
            "description": "The symbol of the asset.",
            "format": "string",
            "nullable": true
          },
          "quantum": {
            "title": "Asset Quantum",
            "type": "string",
            "description": "The quantum of the asset.",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "AssetDtoPaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "title": "Data",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetDto"
            },
            "description": "The data of the response.",
            "nullable": true
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataDto"
          }
        },
        "additionalProperties": false
      },
      "AssetType": {
        "enum": [
          "Eth",
          "Erc20",
          "Erc721",
          "Erc1155",
          "MintableErc20",
          "MintableErc721",
          "MintableErc1155"
        ],
        "type": "string"
      },
...
      "FilterOptions": {
        "enum": [
          "StartsWith",
          "EndsWith",
          "Contains",
          "DoesNotContain",
          "IsEmpty",
          "IsNotEmpty",
          "IsGreaterThan",
          "IsGreaterThanOrEqualTo",
          "IsLessThan",
          "IsLessThanOrEqualTo",
          "IsEqualTo",
          "IsNotEqualTo"
        ],
        "type": "string"
      },
...
      "MetadataDto": {
        "type": "object",
        "properties": {
          "hasNext": {
            "title": "HasNext",
            "type": "boolean",
            "description": "Indicates whether there is a next page available."
          },
          "totalCount": {
            "title": "TotalCount",
            "type": "integer",
            "description": "The total count of results available.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RegisterAssetModel": {
        "required": [
          "name",
          "quantum",
          "symbol",
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/AssetType"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "symbol": {
            "minLength": 1,
            "type": "string"
          },
          "quantum": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "StarkExOperation": {
        "enum": [
          "Deposit",
          "Withdrawal",
          "Mint",
          "MultiTransaction",
          "Transfer",
          "FullWithdrawal",
          "FalseFullWithdrawal",
          "Settlement"
        ],
        "type": "string"
      },
...
      "TransactionDto": {
        "type": "object",
        "properties": {
          "transactionId": {
            "title": "Transaction ID",
            "type": "string",
            "description": "The ID of the transaction.",
            "format": "uuid"
          },
          "operation": {
            "$ref": "#/components/schemas/StarkExOperation"
          },
          "status": {
            "$ref": "#/components/schemas/TransactionStatus"
          },
          "starkExTransaction": {
            "$ref": "#/components/schemas/TransactionModel"
          }
        },
        "additionalProperties": false
      },
      "TransactionDtoPaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "title": "Data",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionDto"
            },
            "description": "The data of the response.",
            "nullable": true
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataDto"
          }
        },
        "additionalProperties": false
      },
...
      "TransactionModel": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "TransactionStatus": {
        "enum": [
          "Created",
          "Streamed",
          "Pending",
          "Confirmed",
          "Reverted",
          "Failed"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://sevensevenbit-dev.eu.auth0.com/oauth/token",
            "scopes": {
              "write:users": "Access user write operations.",
              "read:users": "Access user read operations.",
              "write:assets": "Access asset write operations.",
              "read:assets": "Access asset read operations.",
              "mint:assets": "Access mint operations.",
              "read:vaults": "Access vault read operations.",
              "write:transfers": "Access transfer write operations.",
              "write:settlements": "Access settlement write operations.",
              "read:transactions": "Access tarnsaction read operations."
            }
          }
        }
      }
    }
  },
...
}